+++ quanmiaolightapp-20240801/alibabacloud_quanmiaolightapp20240801/client.py
@@ -3410,6 +3410,8 @@ class Client(OpenApiClient):
             body['snapshotInterval'] = request.snapshot_interval
         if not UtilClient.is_unset(request.split_interval):
             body['splitInterval'] = request.split_interval
+        if not UtilClient.is_unset(request.split_type):
+            body['splitType'] = request.split_type
         if not UtilClient.is_unset(request.task_id):
             body['taskId'] = request.task_id
         if not UtilClient.is_unset(request.text_process_tasks_shrink):
@@ -3503,6 +3505,8 @@ class Client(OpenApiClient):
             body['snapshotInterval'] = request.snapshot_interval
         if not UtilClient.is_unset(request.split_interval):
             body['splitInterval'] = request.split_interval
+        if not UtilClient.is_unset(request.split_type):
+            body['splitType'] = request.split_type
         if not UtilClient.is_unset(request.task_id):
             body['taskId'] = request.task_id
         if not UtilClient.is_unset(request.text_process_tasks_shrink):
@@ -4082,6 +4086,8 @@ class Client(OpenApiClient):
             body['snapshotInterval'] = request.snapshot_interval
         if not UtilClient.is_unset(request.split_interval):
             body['splitInterval'] = request.split_interval
+        if not UtilClient.is_unset(request.split_type):
+            body['splitType'] = request.split_type
         if not UtilClient.is_unset(request.text_process_tasks_shrink):
             body['textProcessTasks'] = request.text_process_tasks_shrink
         if not UtilClient.is_unset(request.video_caption_info_shrink):
@@ -4173,6 +4179,8 @@ class Client(OpenApiClient):
             body['snapshotInterval'] = request.snapshot_interval
         if not UtilClient.is_unset(request.split_interval):
             body['splitInterval'] = request.split_interval
+        if not UtilClient.is_unset(request.split_type):
+            body['splitType'] = request.split_type
         if not UtilClient.is_unset(request.text_process_tasks_shrink):
             body['textProcessTasks'] = request.text_process_tasks_shrink
         if not UtilClient.is_unset(request.video_caption_info_shrink):